terraform: new nodes for variable validation (not yet implemented)
So far we've handled custom validation rules as a part of the same graph
node that is responsible for deciding the final value for an input
variable, but that's bothersome for two reasons:

  - There are two different pieces of code both trying to implement the
    same rules: the root variable node and the module variable node. Those
    have accidentally diverged in the past.
  - Each graph node can evaluate expressions only in one module's scope,
    and so the module variable node is set up to evaluate in the parent
    module's scope because its purpose is to explicitly pass parent module
    data into the child module. That is why we haven't yet allowed input
    variable validation rules to refer to objects declared in the module
    where the variable is declared.

This commit does not actually yet change any real behavior, because this
new graph node has no "Execute" implementation. The focus here is just to
make sure the graph nodes are inserted into the graph with the needed
edges so that a future commit can move the validation logic into here,
and hopefully eventually also allow the validation rules to refer to
objects declared in the child module.