terraform: Move variable validation into its separate node type
The parent commit added a new node to the graph for each existing input
variable node, but that node didn't do anything yet. Now those new nodes
are actually responsible for checking the variable's validation rules,
instead of that being included in the same graph node that decides the
variable's final value.

This also refines the validation node graph transformer to only insert a
validation node if there's at least one rule to check, which should avoid
bloating the graph for anyone who isn't actually using input variable
validation. Therefore the plan graph builder tests revert back to their
old expected result because that input configuration does not include
any validation rules.

For the moment this just moves around functionality that already existed
and doesn't add anything new, but having the validation handled in a
separate graph node means that we can evaluate the validation rules in the
EvalContext of the module where the variable was declared instead of where
it was defined, and so a future commit will allow -- initially just as an
experiment -- validation rules that can refer to other objects in the
module where the variable is declared, including the values of other input
variables.